Outsource functions Test Deno auto imports?
