+++ supabase/functions/evaluate/evaulators/trueOrFalse.ts
@@ -3,7 +3,7 @@ interface CommonResult {
   isCorrect: boolean;
 }
 
-export const evaluateTrueOrFalse = async (questionId: number, answer: string, supabase: any): Promise<CommonResult> => {
+const evaluateTrueOrFalse = async (questionId: number, answer: string, supabase: any): Promise<CommonResult> => {
   const { data, error } = await supabase
     .from('questions')
     .select('*')
